GH #39 - Running tests when building packages

+++ deb/build-zato.sh
@@ -186,26 +186,17 @@ function install_zato {
     fi
 
     ./install.sh -p ${PY_BINARY}
-    pwd && ls -alh
-    for f in zato-server zato-cy;do
-        pushd $i
-        make run-tests || exit 1
-        popd
-    done
-    pushd zato-sso
-        make sso-test || exit 1
-    popd
+    run_tests_zato || exit 1
 
     find $ZATO_TARGET_DIR/. -name *.pyc -exec rm -f {} \;
     find $ZATO_TARGET_DIR/. ! -perm /004 -exec chmod 644 {} \;
-    rm -f ./code/hotfixman.sh
-    rm -rf ./code/hotfixes
+    [[ -f ./code/hotfixman.sh ]] && rm -f ./code/hotfixman.sh
+    [[ -f ./code/hotfixes ]] && rm -rf ./code/hotfixes
     cd $CURDIR
-
 }
 
 function run_tests_zato {
-    for f in zato-server zato-cy;do
+    for i in zato-server zato-cy;do
         pushd $i
         make run-tests || exit 1
         popd
